fix(): tests (#198)
* fix(): tests

* fix(): tests

* fix(): tests

* fix(): tests

+++ gateways/common/naivewatcher/watcher_test.go
@@ -26,7 +26,7 @@ func (w *WatchableTestFS) Walk(root string, walkFn filepath.WalkFunc) error {
 func (w *WatchableTestFS) GetFileID(fi os.FileInfo) interface{} {
 	stat := fi.Sys().(*syscall.Stat_t)
 	return TestFSID{
-		Device: stat.Dev,
+		Device: int32(stat.Dev),
 		Inode:  stat.Ino,
 	}
 }
